LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/no-args-expected.txt


PASS updateCurrentEntry() must throw if state is not given
